refactor :: move package location
