--- src/main/java/com/hackathon/domain/server/domain/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.hackathon.domain.server.domain.user.repository;
-
-import com.hackathon.domain.server.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    Boolean existsByEmail(String email);
-}
