--- src/main/java/com/hackathon/domain/server/domain/user/request/SignupRequest.java
@@ -1,12 +1,0 @@
-package com.hackathon.domain.server.domain.user.request;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class SignupRequest {
-    private String email;
-    private String password;
-    private String name;
-}
