--- src/main/java/com/hackathon/domain/server/domain/user/request/UserInfoRequest.java
@@ -1,11 +1,0 @@
-package com.hackathon.domain.server.domain.user.request;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class UserInfoRequest {
-    private String email;
-    private String name;
-}
