--- src/main/java/com/hackathon/domain/server/domain/user/security/service/UserDetailsServiceImpl.java
@@ -1,25 +1,0 @@
-package com.hackathon.domain.server.domain.user.security.service;
-
-import com.hackathon.domain.server.domain.user.entity.User;
-import com.hackathon.domain.server.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User userEntity = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("해당 이메일을 찾을 수 없습니다: " + email));
-        return UserDetailsImpl.build(userEntity);
-    }
-}
