--- src/main/java/com/hackathon/domain/server/global/exception/BusinessException.java
@@ -1,11 +1,0 @@
-package com.hackathon.domain.server.global.exception;
-
-import com.hackathon.domain.server.global.exception.error.ErrorProperty;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class BusinessException extends RuntimeException {
-    private final ErrorProperty error;
-}
