--- src/main/java/com/hackathon/domain/server/global/exception/error/ErrorProperty.java
@@ -1,8 +1,0 @@
-package com.hackathon.domain.server.global.exception.error;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorProperty {
-    HttpStatus getStatus();
-    String getMessage();
-}
