+++ src/main/java/com/hackathon/server/domain/user/request/LoginRequest.java
@@ -1,0 +1,11 @@
+package com.hackathon.server.domain.user.request;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class LoginRequest {
+    private String email;
+    private String password;
+}
