+++ src/main/java/com/hackathon/server/domain/user/request/SignupRequest.java
@@ -1,0 +1,12 @@
+package com.hackathon.server.domain.user.request;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class SignupRequest {
+    private String email;
+    private String password;
+    private String name;
+}
