+++ src/main/java/com/hackathon/server/domain/user/service/UserService.java
@@ -1,0 +1,8 @@
+package com.hackathon.server.domain.user.service;
+
+import com.hackathon.server.global.exception.BusinessException;
+import com.hackathon.server.domain.user.request.SignupRequest;
+
+public interface UserService {
+    void registerUser(SignupRequest signupRequest) throws BusinessException;
+}
