+++ src/main/java/com/hackathon/server/global/exception/BusinessException.java
@@ -1,0 +1,11 @@
+package com.hackathon.server.global.exception;
+
+import com.hackathon.server.global.exception.error.ErrorProperty;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class BusinessException extends RuntimeException {
+    private final ErrorProperty error;
+}
