+++ src/main/java/com/hackathon/server/global/exception/error/ErrorProperty.java
@@ -1,0 +1,8 @@
+package com.hackathon.server.global.exception.error;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorProperty {
+    HttpStatus getStatus();
+    String getMessage();
+}
